+++ src/index.js
@@ -30,11 +30,30 @@ if (
   typeof navigator !== 'undefined' &&
   navigator.product === 'ReactNative'
 ) {
+  // eslint-disable-next-line no-console
   console.warn(
     "It looks like you've imported 'styled-components' on React Native.\n" +
       "Perhaps you're looking to import 'styled-components/native'?\n" +
       'Read more about this at https://www.styled-components.com/docs/basics#react-native'
   )
+}
+
+/* Warning if there are several instances of styled-components */
+if (process.env.NODE_ENV !== 'production' && typeof window !== 'undefined') {
+  window['__styled-components-init__'] =
+    window['__styled-components-init__'] || 0
+
+  if (window['__styled-components-init__'] === 1) {
+    // eslint-disable-next-line no-console
+    console.warn(
+      "It looks like there are several instances of 'styled-components' initialized in this application. " +
+        'This may cause dynamic styles not rendering properly, errors happening during rehydration process ' +
+        'and makes you application bigger without a good reason.\n\n' +
+        'See https://s-c.sh/2BAXzed for more info.'
+    )
+  }
+
+  window['__styled-components-init__'] += 1
 }
 
 /* Instantiate singletons */