+++ src/models/ServerStyleSheet.js
@@ -133,12 +133,10 @@ class ServerTag implements Tag {
 export default class ServerStyleSheet {
   closed: boolean
   instance: StyleSheet
-  isStreaming: boolean
-  lastIndex: number
 
   constructor() {
     this.instance = StyleSheet.clone(StyleSheet.instance)
-    this.isStreaming = false
+    this.instance.isStreaming = false
   }
 
   collectStyles(children: any) {
@@ -182,18 +180,18 @@ export default class ServerStyleSheet {
       // $FlowFixMe
       ourStream._read = () => {}
 
-      this.isStreaming = true
-      this.lastIndex = 0
+      this.instance.isStreaming = true
 
       readableStream.on('data', chunk => {
         ourStream.push(
-          this.instance.tags
-            .slice(this.lastIndex)
-            .map(tag => tag.toHTML())
-            .join('') + chunk
-        )
+          this.instance.tags.reduce((html, tag) => {
+            if (!tag.isSealed()) {
+              html += tag.toHTML() // eslint-disable-line no-param-reassign
+            }
 
-        this.lastIndex = this.instance.tags.length - 1
+            return html
+          }, '') + chunk
+        )
       })
 
       readableStream.on('end', () => {